// src/app/model/employee.model.ts
export interface Employee {
    id: number;
    name: string;
    gender: string;
    email: string;
    phone: number;
    department: string;
    image: string;
    birthDate: string; 
    startDate: string; 
    salary: string; 
    homeOffice: string; 
    education: string; 
    jobTitle: string; 
    description: string;
}